feat(scw): merge session and sampler
