+++ cirq-scaleway/README.rst
@@ -25,6 +25,34 @@ Note, that this will install both **cirq-scaleway** and **cirq-core**.
 
 To get all the optional modules installed, you'll have to use `pip install cirq` or `pip install cirq~=1.0.dev` for the pre-release version.
 
+Getting started
+---------------
+
+In the most simple way, here the code to use **cirq-scaleway**:
+
+>>> import cirq
+>>> from cirq_scaleway import ScalewayQuantumService
+
+>>> service = ScalewayQuantumService(
+    project_id="<your-scaleway-project-id>", secret_key="<your-scaleway-secret-key>"
+)
+
+>>> # Get and display all provided (real or simulated) devices
+>>> devices = service.devices(min_num_qubits=34)
+>>> print(devices)
+
+>>> # Get a specific device by its name
+>>> qsim_simulator = service.device(device="qsim_simulation_c64m512")
+
+>>> # Create a device session and run a circuit against it
+>>> with qsim_simulator.create_session():
+>>>   qubit = cirq.GridQubit(0, 0)
+>>>   circuit = cirq.Circuit(cirq.X(qubit) ** 0.5, cirq.measure(qubit, key='m'))
+
+>>>   # Run the circuit on the device
+>>>   result = sampler.run(circuit)
+>>>   print(result)
+
 Development
 ------------
